Implement accumulative kills per monster type in playerstats cmd

+++ src/r_segs.c
@@ -312,9 +312,8 @@ void R_RenderMaskedSegRange(drawseg_t *ds, int x1, int x2)
                 int64_t     t = ((int64_t)centeryfrac << FRACBITS)
                                 - (int64_t)dc_texturemid * spryscale;
 
-                if (t + (int64_t)texheight * spryscale < 0
-                    || t > (int64_t)(SCREENHEIGHT << FRACBITS) * 2)
-                    continue;                       // skip if the texture is out of screen's range
+                if (t + (int64_t)texheight * spryscale < 0 || t > (SCREENHEIGHT << FRACBITS) * 2)
+                    continue;                    // skip if the texture is out of screen's range
 
                 sprtopscreen = (int64_t)(t >> FRACBITS);
             }
